integrating tenant aspect for topics, moving tenant aspect to proper package
