+++ pom.xml
@@ -214,6 +214,18 @@
       <scope>test</scope>
     </dependency>
 
+
+    <dependency>
+      <groupId>org.mariadb.jdbc</groupId>
+      <artifactId>mariadb-java-client</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>com.h2database</groupId>
+      <artifactId>h2</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
   <repositories>
@@ -456,15 +468,6 @@
           </execution>
         </executions>
       </plugin>
-      <!-- Liquibase Maven plugin -->
-<!--      <plugin>-->
-<!--        <groupId>org.liquibase</groupId>-->
-<!--        <artifactId>liquibase-maven-plugin</artifactId>-->
-<!--        <version>4.1.1</version>-->
-<!--        <configuration>-->
-<!--          <propertyFile>src/main/resources/liquibase.properties</propertyFile>-->
-<!--        </configuration>-->
-<!--      </plugin>-->
     </plugins>
   </build>
   <profiles>
