+++ src/test/java/de/caritas/cob/consultingtypeservice/api/consultingtypes/ConsultingTypeGroupTenantAwareRepositoryImplTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
 import de.caritas.cob.consultingtypeservice.api.model.ConsultingTypeEntity;
-import de.caritas.cob.consultingtypeservice.api.service.tenant.TenantContext;
+import de.caritas.cob.consultingtypeservice.api.tenant.TenantContext;
 import de.caritas.cob.consultingtypeservice.schemas.model.ConsultingType;
 import java.util.Arrays;
 import java.util.Collections;
